# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: New Model Bread - Bakery Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Soft delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Branches
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales
      parameters:
        - name: branchId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: startDate
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Record a sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleBody"
      responses:
        "201":
          description: Created sale with receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "400":
          description: Insufficient stock or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get sale details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/daily-summary:
    get:
      operationId: getDailySalesSummary
      tags: [sales]
      summary: Get daily sales summary
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: branchId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Daily sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySalesSummary"

  # Reports/Dashboard
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # Auth schemas
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [managing_director, manager, receptionist, production_staff]
        branchId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        deletedAt:
          type: ["string", "null"]
      required:
        - id
        - username
        - fullName
        - role
        - isActive
        - createdAt
        - updatedAt
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        fullName:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [managing_director, manager, receptionist, production_staff]
        branchId:
          type: ["integer", "null"]
      required:
        - username
        - password
        - fullName
        - role
    UpdateUserBody:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        branchId:
          type: ["integer", "null"]
        isActive:
          type: ["boolean", "null"]
        password:
          type: ["string", "null"]

    # Branch schemas
...
    Sale:
      type: object
      properties:
        id:
          type: integer
        receiptNumber:
          type: string
        breadType:
          type: string
        quantity:
          type: integer
        pricePerUnit:
          type: number
        totalAmount:
          type: number
        costAmount:
          type: number
        profitAmount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, transfer]
        cashierId:
          type: integer
        cashierName:
          type: string
        branchId:
          type: integer
        branchName:
          type: string
        notes:
          type: ["string", "null"]
        saleDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - receiptNumber
        - breadType
        - quantity
        - pricePerUnit
        - totalAmount
        - costAmount
        - profitAmount
        - paymentMethod
        - cashierId
        - cashierName
        - branchId
        - branchName
        - saleDate
        - createdAt
    CreateSaleBody:
      type: object
      properties:
        breadType:
          type: string
        quantity:
          type: integer
        pricePerUnit:
          type: number
        paymentMethod:
          type: string
          enum: [cash, transfer]
        branchId:
          type: integer
        notes:
          type: ["string", "null"]
      required:
        - breadType
        - quantity
        - pricePerUnit
        - paymentMethod
        - branchId
    DailySalesSummary:
      type: object
      properties:
        date:
          type: string
        totalSales:
          type: integer
        totalRevenue:
          type: number
        totalProfit:
          type: number
        totalCost:
          type: number
        cashSales:
          type: number
        transferSales:
          type: number
        breadTypes:
          type: array
          items:
            $ref: "#/components/schemas/BreadTypeSummary"
      required:
        - date
        - totalSales
        - totalRevenue
        - totalProfit
        - totalCost
        - cashSales
        - transferSales
        - breadTypes
    BreadTypeSummary:
      type: object
      properties:
        breadType:
          type: string
        quantity:
          type: integer
        revenue:
          type: number
      required:
        - breadType
        - quantity
        - revenue

    # Dashboard schemas
...