# lib/api-zod/src/generated/types/createSaleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * New Model Bread - Bakery Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSaleBodyPaymentMethod } from "./createSaleBodyPaymentMethod";

export interface CreateSaleBody {
  breadType: string;
  quantity: number;
  pricePerUnit: number;
  paymentMethod: CreateSaleBodyPaymentMethod;
  branchId: number;
  /** @nullable */
  notes?: string | null;
}